requirements.txt
asgiref==3.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.6
psycopg2==2.9.10
PyJWT==2.7.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
